% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_btc.R
\name{read.btc}
\alias{read.btc}
\title{Read bitcoinchart data}
\usage{
read.btc(..., folder = ".")
}
\arguments{
\item{folder}{path to folder where the data is saved}

\item{exc}{Exchange}

\item{cur}{Fiat money currency}

\item{file}{File name which links directly to saved files (use either a exc, cur pair **or** file)}
}
\description{
Read bitcoinchart data
}
